Finish largest part of project rewrite/restructure
All backends have been rewritten so they are compatible with the new interface. The backends have been rewritten from scratch, so as to ensure that they are compatible with the latest versions of the SDKs. There are a few TODOs left though.

The Windows and Linux backends for the MasterKeys keyboards have been tested for the MasterKeys Pro L RGB.